+++ scylla/src/session/session_test.rs
@@ -7,12 +7,12 @@ use crate::execution::retries::{QueryInfo, RetryDecision, RetryPolicy, RetrySess
 use crate::execution::tracing::TracingInfo;
 use crate::prepared_statement::PreparedStatement;
 use crate::query::Query;
+use crate::routing::partitioner::{
+    calculate_token_for_partition_key, Murmur3Partitioner, Partitioner, PartitionerName,
+};
 use crate::routing::Token;
 use crate::session::Session;
 use crate::statement::Consistency;
-use crate::transport::partitioner::{
-    calculate_token_for_partition_key, Murmur3Partitioner, Partitioner, PartitionerName,
-};
 use crate::utils::test_utils::{
     create_new_session_builder, scylla_supports_tablets, setup_tracing, supports_feature,
     unique_keyspace_name, PerformDDL,
