+++ scylla/src/transport/mod.rs
@@ -1,6 +1,5 @@
 pub mod legacy_query_result;
 pub mod locator;
-pub mod partitioner;
 pub mod query_result;
 
 pub use crate::frame::Authenticator;
